# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "後端專案",
        "description": "\n *     後端專案專用的 ＡＰＩ 文件,\n *     請先透過 auth 分類中的 login 登入拿到 token\n *     在到右邊的 Authorize 按鈕中中設定，方可使用所有有鎖頭的 api\n *     然後有鎖頭的 ＡＰＩ 要在 http header 帶 token 進去\n *     ",
        "contact": {
            "name": "stephen chen",
            "email": "tasb00429@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/admins/banners": {
            "get": {
                "tags": [
                    "Banner"
                ],
                "summary": "把 Banner 資料列出來",
                "operationId": "93eddce59df7e1b5515b99dfbd154b91",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "admin 的 page ( 沒有帶參數就全部吐回來 )",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "admin 的 per_page ( 沒有帶參數就全部吐回來 )",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Banner"
                ],
                "summary": "新增 Banner, permission 也一並傳近來新增",
                "operationId": "2aadb84170514ef709c42e91d46f22e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "描述",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "狀態",
                                        "type": "boolean"
                                    },
                                    "permission_ids": {
                                        "description": "permissions 的 ID",
                                        "type": "array",
                                        "items": {
                                            "description": "permissions 的 id",
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admins/banners/{id}": {
            "get": {
                "tags": [
                    "Banner"
                ],
                "summary": "查看一筆 Banner, 會把 permissions 一起回傳",
                "operationId": "c3b9372a9f0d3f1b851a5b6f90b7e363",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "banner 的 id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Banner"
                ],
                "summary": "修改一筆 Banner, permission 也一並傳近來一起修改",
                "operationId": "75a6ed13abae9f73d9eeaa925e8ba319",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "banner 的 id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Updated name of the pet",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "描述",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "狀態",
                                        "type": "boolean"
                                    },
                                    "permission_ids": {
                                        "description": "permissions 的 ID",
                                        "type": "array",
                                        "items": {
                                            "description": "permissions 的 id",
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Banner"
                ],
                "summary": "刪除一筆 Banner",
                "operationId": "6719aabd98249a109efa89073086733f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "banner 的 id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admins/permissions": {
            "get": {
                "tags": [
                    "Permission"
                ],
                "summary": "把 Permission 資料列出來",
                "operationId": "b48c083bdf2f2599353d2dd2da392539",
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}